Moved memberCount to /courses instead of /projects
