+++ backend/app/src/main/java/com/ugent/pidgeon/model/json/CourseWithRelationJson.java
@@ -1,12 +1,16 @@
 package com.ugent.pidgeon.model.json;
 
 import com.ugent.pidgeon.postgre.models.types.CourseRelation;
+import java.time.OffsetDateTime;
 
 public class CourseWithRelationJson {
     private Long courseId;
     private String url;
     private String name;
-    private Boolean archived;
+    private OffsetDateTime archivedAt;
+    private CourseRelation relation;
+    private Integer memberCount;
+
 
 
     public String getUrl() {
@@ -17,20 +21,17 @@ public class CourseWithRelationJson {
         this.url = url;
     }
 
-    public CourseWithRelationJson(Boolean archived) {
-      this.archived = archived;
-    }
 
     public CourseWithRelationJson(String url, CourseRelation relation, String name, Long courseId,
-        Boolean archived) {
+        OffsetDateTime archivedAt, Integer memberCount) {
         this.url = url;
         this.relation = relation;
         this.name = name;
         this.courseId = courseId;
-        this.archived = archived;
+        this.archivedAt = archivedAt;
+        this.memberCount = memberCount;
     }
 
-    private CourseRelation relation;
 
     public CourseRelation getRelation() {
         return relation;
@@ -56,11 +57,19 @@ public class CourseWithRelationJson {
         this.courseId = id;
     }
 
-    public Boolean getArchived() {
-        return archived;
+    public OffsetDateTime getArchivedAt() {
+        return archivedAt;
     }
 
-    public void setArchived(Boolean archived) {
-        this.archived = archived;
+    public void setArchivedAt(OffsetDateTime archivedAt) {
+        this.archivedAt = archivedAt;
+    }
+
+    public Integer getMemberCount() {
+        return memberCount;
+    }
+
+    public void setMemberCount(Integer memberCount) {
+        this.memberCount = memberCount;
     }
 }
